seed/requirements.txt
#
# __Seed builder__
#   AUTO_GENERATED (Read only)
#   Modify via builder
#

wheel==0.41.2
apipkg==3.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
coverage==7.3.1
Django==4.2.5
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-filter==23.2
djangoql==0.17.1
django-import-export==3.2.0
dj-rest-auth==4.0.1
django-storages==1.14
djangorestframework==3.14.0
graphene-django==3.1.5
gunicorn==21.2.0
psycopg2-binary==2.9.7
pylint==2.17.5
redis==5.0.0
requests==2.31.0
setuptools==68.2.0
six==1.16.0